Fix int8 -> int4

+++ etl-aro/etl/financial/create_financials.sql
@@ -411,7 +411,7 @@ DROP TABLE IF EXISTS "financial"."fiber_summary_cost";
 create table financial.fiber_summary_cost (
 	id bigserial primary key,
 	
-	network_cost_code_id int8 references financial.network_cost_code,
+	network_cost_code_id int4 references financial.network_cost_code,
 	network_report_id int8 not null references financial.network_report on delete cascade,
 		
 	length_meters double precision,
